+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.2]() - 2025-06-21
+### Fixed
+- ENVs overwrite Settings in database properly now
+- Library providers sync all watch history on first run, settings.app.recent_limit after.
+
 ## [1.1.1]() - 2025-06-20
 ### Fixed
 - Fixed issue with favorite returning None in Plex Provider causing watch history sync to fail.